+++ sanity-studio/schemaTypes/index.ts
@@ -1,0 +1,21 @@
+import { blockContent }  from './blockContent'
+import { proyecto }      from './proyecto'
+import { articulo }      from './articulo'
+import { servicio }      from './servicio'
+import { producto }      from './producto'
+import { paginaCliente } from './paginaCliente'
+import { testimonio }    from './testimonio'
+import { faqItem }       from './faqItem'
+import { pedido }        from './pedido'
+
+export const schemaTypes = [
+  blockContent,
+  proyecto,
+  articulo,
+  servicio,
+  producto,
+  paginaCliente,
+  testimonio,
+  faqItem,
+  pedido,
+]